02_AI_Agents/jupyter-distributed-kernel/requirements.txt
gradio
aiohttp
python-dotenv
groq
websockets
fastapi
uvicorn
